chore: Extract all of the type declarations into a central location. Move all of the type declarations in separate files under src/types directory. Redefine the entities to be expressed more idiomatically.
