+++ bEnd/Kanban/deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "dev": "deno run --watch --allow-net --allow-read --allow-write --allow-env main.ts",
+    "dev": "deno run --watch --allow-net --allow-read --allow-write --allow-env --env-file=.env main.ts",
     "generate": "deno run -A --node-modules-dir npm:drizzle-kit generate",
     "migrate": "deno run -A --node-modules-dir npm:drizzle-kit migrate"
   },
