--- bEnd/Kanban/src/https/types/data-types.ts
@@ -1,17 +1,0 @@
-import {
-  Board,
-  Column,
-  Comment,
-  Section,
-  Task,
-  TaskToDo,
-} from "../../websockets/types/entities.ts";
-
-export type EntityCollection = {
-  sections: Section[];
-  boards: Board[];
-  columns: Column[];
-  tasks: Task[];
-  taskToDos: TaskToDo[];
-  comments: Comment[];
-};
