+++ bEnd/Kanban/src/https/utils/dtos.ts
@@ -1,9 +1,9 @@
-import { User } from "../types/userTypes.ts";
+import { User } from "../../types/entities.ts";
 
 export const userDto = (user: User): Partial<User> => ({
-    id: user.id,
-    firstName: user.firstName,
-    lastName: user.lastName,
-    email: user.email,
-    role: user.role,
+  id: user.id,
+  firstName: user.firstName,
+  lastName: user.lastName,
+  email: user.email,
+  role: user.role,
 });
