+++ bEnd/Kanban/src/https/utils/jsonWebToken.ts
@@ -1,25 +1,25 @@
 import { create, getNumericDate, type Payload, verify } from "jwt";
-import type { User } from "../types/userTypes.ts";
+import type { User } from "../../types/entities.ts";
 
 export async function createJWT(
-    user: User,
-    key: CryptoKey,
-    exp: number,
+  user: User,
+  key: CryptoKey,
+  exp: number,
 ): Promise<string> {
-    const jwt = await create(
-        { alg: "HS512", typ: "JWT" },
-        {
-            id: user.id,
-            email: user.email,
-            fName: user.firstName,
-            lName: user.lastName,
-            exp: getNumericDate(60 * exp),
-        },
-        key,
-    );
-    return jwt;
+  const jwt = await create(
+    { alg: "HS512", typ: "JWT" },
+    {
+      id: user.id,
+      email: user.email,
+      fName: user.firstName,
+      lName: user.lastName,
+      exp: getNumericDate(60 * exp),
+    },
+    key,
+  );
+  return jwt;
 }
 
 export async function verifyJWT(jwt: string, key: CryptoKey): Promise<Payload> {
-    return await verify(jwt, key);
+  return await verify(jwt, key);
 }
