+++ bEnd/Kanban/src/types/errormap.ts
@@ -1,0 +1,10 @@
+export const ErrorCodeMap = {
+  // Bad Request
+  400: 1002,
+  // Unauthorized
+  401: 3000,
+  // Unsuported Data
+  415: 1003,
+  // Conflict / Policy Violation
+  409: 1008,
+};
