+++ bEnd/Kanban/src/types/extentions/ws.d.ts
@@ -1,0 +1,10 @@
+import { Payload } from "jwt";
+
+declare global {
+  interface WebSocket {
+    _token: string;
+    _user: Omit<Payload, "exp">;
+  }
+}
+
+export {};
