+++ bEnd/Kanban/src/websockets/services/task.service.ts
@@ -1,6 +1,6 @@
-import { Task } from "../types/entities.ts";
+import { Task } from "../../types/entities.ts";
 import { ApiError } from "../../errors/apiErrors.ts";
-import { MessageMap } from "../types/zod/inbound.ts";
+import { MessageMap } from "../../types/validation.ts";
 import { DatabaseError } from "../../errors/databaseErrors.ts";
 import { _deleteTask, _insertTask, _updateTask } from "../../database/db.ts";
