--- bEnd/Kanban/src/websockets/typings/ws.d.ts
@@ -1,10 +1,0 @@
-import { Payload } from "jwt";
-
-declare global {
-    interface WebSocket {
-        _token: string;
-        _user: Omit<Payload, "exp">;
-    }
-}
-
-export {};
